Refactor Ballerina project target lookup and use for LS
